MoE a16wmix: share the A and B operand paths between gemm1 and gemm2
gemm1 and gemm2 contract the same way -- N-major preshuffled W, per-lane scalar
scale gather, upconvert to v8bf16 MMA fragments, A staged through LDS and read
back on the CK sub-lane map -- and differ only in what N and K mean (stage1:
N=2*inter_dim gate|up, K=model_dim; stage2: N=model_dim, K=inter_dim).

Both files carried their own copy. Compared with comments stripped, these were
already character-identical: load_b_raw / load_b_raw_int4 / load_b_raw_bf16,
load_b_scale / load_b_scale_int4, _a_col_bytes_for_ku. upconvert_b differed only
by a local name (p vs pp), and the W/W-scale layout + buffer-resource setup
feeding them was duplicated too. The copies had ALREADY drifted in their
comments, which is the real cost: a reader can't tell whether they've also
drifted in behaviour.

Extract into utils:
  make_b_loader() -> _BLoader   W layouts, buffer resources, copy atoms and the
                                raw/scale/upconvert closures, keyed on w_dtype
                                (fp4 mxfp4 / int4 / bf16)
  make_a_loader() -> _ALoader   A-LDS view + v8bf16 fragment reader; `tiles` is
                                exposed because the global->LDS DMA writes
                                through the same view

The genuine stage differences become named parameters rather than divergent
copies: make_a_loader takes `swizzle` (gemm2 XOR-swizzles A-LDS, gemm1 is
linear) and `k_grp_base_bytes`/`A_SLOT_BYTES` (gemm1's k_wave group +
double-buffer slot; None for gemm2's single unslotted region).
load_b_scale_int4 now takes scale_expert_base as an argument instead of reaching
for a late-bound enclosing local defined 100 lines below its definition.

Also rename mma_bf16 -> _mma_bf16: every other helper in utils is underscore-
private, including its own _bf16_frag8/_bf16_frag4, and there is no __all__ or
external consumer.

Verified BYTE-EXACT: 168 configs (stage1+stage2 x gfx950/gfx942 x fp4/int4/bf16
x tile_m 16/64 x tile_n x k_wave 1/2/4 x b_nt 0/2 x persist) compiled cold into
a scratch cache before and after; all 118 that produce a kernel have identical
gpu.binary ELF sha256, and the 30 LDS-overflow aborts + 20 gfx942-fp4 compile
rejects are unchanged too (so the refactor can't have silently turned an error
into a kernel). Re-run after each extraction step.
op_tests/test_moe_2stage.py --csv-filter wint4 abf16_wfp4: 88/88, err
1.42e-05..1.62e-05 (threshold 0.01).
