+++ aiter/ops/flydsl/kernels/moe_2stage_a16wmix/gemm1.py
@@ -12,23 +12,20 @@ from flydsl.expr.typing import Vector as Vec
 
 from aiter.ops.flydsl.kernels import buffer_ops
 from aiter.ops.flydsl.kernels.act import _silu_mul_batch, _situ_mul_batch
-from aiter.ops.flydsl.kernels.layout_utils import crd2idx
 from aiter.ops.flydsl.kernels.tensor_shim import _to_raw as _raw
 
 from .utils import (
     A16WI4_GROUP_SIZE,
     _a16w4_swizzle_xor16,
-    _buffer_i32_scalar_read,
-    _e8m0_byte_to_f32,
     _gep,
     _global_i32_at,
-    _global_i32_buffer_tiles,
     _global_i32_buffer_view,
-    _int4_nibble_to_bf16x8,
     _lds_ptr3,
+    _mma_bf16,
     _udiv,
     _umod,
-    mma_bf16,
+    make_a_loader,
+    make_b_loader,
 )
 
 # =============================================================================
@@ -86,7 +83,6 @@ def _gemm1_body_a16w4(
     LDS_STRIDE = TILE_K  # bf16 elems per LDS row (pad_k=0, LDS128)
     m_repeat = BM // 16
     k_unroll = KH_TILE_BYTES // 64  # bf16 8-per-lane K micro-steps per K-tile
-    _k0_count = TILE_K // 128
     # Wave partition num_n_waves x k_wave. k_wave=1: 4 waves split TILE_N (TILE_N/4 each).
     # k_wave>1 (aiter intra-block slice-K): each wave does a K-slice (klen=K/k_wave) of a
     # wider N-slice; partials LDS-reduced across k-group peers before epilogue.
@@ -115,47 +111,6 @@ def _gemm1_body_a16w4(
     _A_GRP_BYTES = A_LDS_STAGES * A_SLOT_BYTES
     NUM_N_BLOCKS = INTER // TILE_N
 
-    # W (mxfp4) preshuffle layout (aiter make_preshuffle_b_layout, N-major, fp4 bytes):
-    #   shape (N_OUT/16, (K/2)/64, klane=4, nlane=16, kpack=16), strides below.
-    K_HALF = K // 2
-    bl_k0 = K_HALF // 64
-    bl_stride_klane = 256
-    bl_stride_k0 = 1024
-    bl_stride_n0 = bl_k0 * bl_stride_k0
-    layout_b = fx.make_layout(
-        (N_OUT // 16, bl_k0, 4, 16, 16),
-        (bl_stride_n0, bl_stride_k0, bl_stride_klane, 16, 1),
-    )
-    # int4 W: OLD-kernel preshuffle, kpack=8 (K16/klane slot), strides (n0,512,128,8,1);
-    # one MFMA-K32 = two klane slots 128B apart, bytes interleaved-by-4 (byte j = K_j|K_{j+4}<<4).
-    if const_expr(_is_int4):
-        i4l_k0 = K // 64
-        i4l_stride_klane = 128
-        i4l_stride_k0 = 512
-        i4l_stride_n0 = i4l_k0 * i4l_stride_k0
-        layout_b_int4 = fx.make_layout(
-            (N_OUT // 16, i4l_k0, 4, 16, 8),
-            (i4l_stride_n0, i4l_stride_k0, i4l_stride_klane, 8, 1),
-        )
-    # W (raw bf16) preshuffle layout (N-major == shuffle_weight (16,16)), bf16-elem units:
-    #   shape (N_OUT/16, K/32, 4, 16, 8). One kpack = 8 bf16 = ONE MFMA K32 B fragment
-    #   (no upconvert). K reindexed to the fp4 (klane_hw, ku)->K order (see load_b_raw_bf16).
-    bfl_k0 = K // 32
-    bfl_stride_klane = 128
-    bfl_stride_k0 = 512
-    bfl_stride_n0 = bfl_k0 * bfl_stride_k0
-    layout_b_bf16 = fx.make_layout(
-        (N_OUT // 16, bfl_k0, 4, 16, 8),
-        (bfl_stride_n0, bfl_stride_k0, bfl_stride_klane, 8, 1),
-    )
-    # B-scale preshuffle layout (make_preshuffle_scale_layout, e8m0 u8, per-1x32):
-    #   K padded to 256 mult. shape (N_OUT/32, c_k1, 4, 16), strides below.
-    scale_k_padded = ((K + 255) // 256) * 256
-    sc_k1 = ((scale_k_padded // 32) // 4) // 2
-    sc_stride_klane = 16
-    sc_stride_k0 = 64
-    sc_stride_n0 = sc_k1 * sc_stride_k0
-
     # a16wi4 groupwise scale: bf16 pairs, layout (E, N, G//2, 2).
     _num_groups = K // A16WI4_GROUP_SIZE
     _g_half = _num_groups // 2
@@ -172,42 +127,23 @@ def _gemm1_body_a16w4(
     expert_off = e * fx.Int32(N_OUT)
     inter_i32 = fx.Int32(INTER)
 
-    # ---- buffer resources -----------------------------------------------------
-    # bf16 W [E, N_OUT, K] whole-tensor extent overflows the 32-bit num_records/i32
-    # byte-offset at large E (E896: 6.6GB): fold the per-expert base into the i64
-    # resource addr and index within the expert. mxfp4/int4 keep the whole-tensor path.
-    if const_expr(_is_bf16):
-        _w_per_expert_bytes = N_OUT * (K * 2)
-        w_base_i64 = fx.Int64(arg_bq) + fx.Int64(e) * fx.Int64(_w_per_expert_bytes)
-        w_tiles = _global_i32_buffer_tiles(
-            w_base_i64, min(_w_per_expert_bytes, 0xFFFFFFFF), 4
-        )
-    else:
-        _w_bytes = NE * N_OUT * K_HALF
-        w_tiles = _global_i32_buffer_tiles(arg_bq, min(_w_bytes, 0xFFFFFFFF), 4)
-    w_copy_atom = fx.make_copy_atom(fx.rocdl.BufferCopy128b(b_cache_mod), fx.Int32)
-    w_reg_lay = fx.make_layout(4, 1)
-    if const_expr(_is_int4):
-        # OLD int4 layout: one klane slot = 8 bytes = 2 i32 -> dwordx2 (64b) loads. A
-        # port MFMA-K32 block is two such slots 128 B apart (see load_b_raw int4 branch).
-        w_copy_atom64 = fx.make_copy_atom(fx.rocdl.BufferCopy64b(b_cache_mod), fx.Int32)
-        w_reg_lay2 = fx.make_layout(2, 1)
-        w_tiles8 = _global_i32_buffer_tiles(arg_bq, min(_w_bytes, 0xFFFFFFFF), 2)
-    if _is_int4:
-        # int4 groupwise scale buffer (E, N_OUT, G//2, 2) bf16 -> G//2 dwords per N.
-        _sw_bytes = NE * N_OUT * _g_half * 4
-    else:
-        _sw_bytes = NE * N_OUT * (scale_k_padded // 32)
-    # bf16 W has no scale buffer; sw_tiles unused (arg_bscale is a dummy pointer). Scale
-    # is a per-lane scalar e8m0/bf16-pair gather: a make_buffer_tensor 1-dword tiles view
-    # + BufferCopy32b scalar read (buffer_load_dword, OOB-clamped) replaces the raw
-    # create_buffer_resource_from_addr + buffer_load.
-    sw_tiles = (
-        None
-        if _is_bf16
-        else _global_i32_buffer_tiles(arg_bscale, min(_sw_bytes, 0xFFFFFFFF), 1)
+    # ---- B (weight) operand path: layouts + buffer resources + load closures ----
+    # Shared verbatim with gemm2 (see utils.make_b_loader); stage1's N is the gate|up
+    # 2*inter_dim and its K is model_dim.
+    b_loader = make_b_loader(
+        arg_bq,
+        arg_bscale,
+        N_OUT=N_OUT,
+        K=K,
+        NE=NE,
+        e=e,
+        lane_div_16=lane_div_16,
+        lane_mod_16=lane_mod_16,
+        TILE_K=TILE_K,
+        w_dtype=w_dtype,
+        b_cache_mod=b_cache_mod,
+        use_k16=use_k16,
     )
-    sw_read_atom = fx.make_copy_atom(fx.rocdl.BufferCopy32b(0), fx.Int32)
     # Intermediate [sorted_size, inter] bf16: num_records = cumsum0*inter*2, so masked
     # (clamped) stores land OOB. KEPT RAW: the output resource + masked buffer_store need a
     # dynamic (runtime cumsum0) num_records and per-store predication; the fx.copy layout
@@ -302,197 +238,31 @@ def _gemm1_body_a16w4(
                 fx.copy(
                     x_lds_store_atom,
                     r,
-                    fx.slice(s_x_i32x4_tiles, (None, lds_byte // fx.Int32(16))),
+                    fx.slice(a_loader.tiles, (None, lds_byte // fx.Int32(16))),
                 )
             else:
                 fx.copy(
                     x_dma_atom,
                     fx.slice(x_dma_tiles4, (None, global_byte // fx.Int32(16))),
-                    fx.slice(s_x_i32x4_tiles, (None, lds_byte // fx.Int32(16))),
+                    fx.slice(a_loader.tiles, (None, lds_byte // fx.Int32(16))),
                 )
 
     # ---- A LDS read (CK sub-lane): lane L covers K[L*32..L*32+31] --------------
     # Each (mi, ku) reads 8 bf16 (one ds_read_b128) -> v8bf16 A operand.
-    row_a_lds = lane_mod_16
-    col_base_bytes_L = lane_div_16 * fx.Int32(64)  # 32 bf16 * 2 B
-    s_x_i32_flat = fx.make_view(
-        fx.recast_iter(fx.Int32, lds_raw_ptr),
-        fx.make_layout(k_wave * A_LDS_STAGES * BM * LDS_STRIDE // 2, 1),
+    # Shared with gemm2 (see utils.make_a_loader); stage1's A-LDS is carved into
+    # k_wave groups x 2 pipeline slots, and is NOT XOR-swizzled (its DMA source is
+    # already conflict-free).
+    a_loader = make_a_loader(
+        lds_raw_ptr,
+        num_i32=k_wave * A_LDS_STAGES * BM * LDS_STRIDE // 2,
+        KH_TILE_BYTES=KH_TILE_BYTES,
+        k_blocks16=k_blocks16,
+        lane_div_16=lane_div_16,
+        lane_mod_16=lane_mod_16,
+        swizzle=False,
+        k_grp_base_bytes=k_grp_base_bytes,
+        A_SLOT_BYTES=A_SLOT_BYTES,
     )
-    s_x_i32x4_tiles = fx.logical_divide(s_x_i32_flat, fx.make_layout(4, 1))
-    a_copy_atom = fx.make_copy_atom(fx.UniversalCopy128b(), fx.Int32)
-
-    def _a_col_bytes_for_ku(ku):
-        _k0_blk = ku // 4
-        _ku_in = ku % 4
-        return col_base_bytes_L + fx.Int32(_ku_in * 16 + _k0_blk * 256)
-
-    def lds_load_a(mi, ku, slot=0):
-        row = row_a_lds + fx.Int32(mi * 16)
-        col_swz_bytes = _a16w4_swizzle_xor16(
-            row, _a_col_bytes_for_ku(ku), fx.Int32(k_blocks16)
-        )
-        # byte offset within this k-group's A-LDS slot -> 16-byte tile index.
-        byte_off = (
-            k_grp_base_bytes
-            + fx.Int32(slot * A_SLOT_BYTES)
-            + row * fx.Int32(KH_TILE_BYTES)
-            + col_swz_bytes
-        )
-        r = fx.make_rmem_tensor(fx.make_layout(4, 1), fx.Int32)
-        fx.copy_atom_call(
-            a_copy_atom, fx.slice(s_x_i32x4_tiles, (None, byte_off // fx.Int32(16))), r
-        )
-        return fx.Vector(fx.memref_load_vec(r)).bitcast(fx.BFloat16)  # v8bf16
-
-    # ---- B (mxfp4 W) raw load: dwordx4 -> v4i32 (8 fp4 per i32) ----------------
-    def load_b_raw(base_k, n_blk, n_intra):
-        # raw[k0][j] = i32 holding 8 fp4 for K micro-step (k0*4 + j).
-        raw = []
-        for k0i in range_constexpr(_k0_count):
-            k0 = (base_k + fx.Int32(k0i * 128)) // fx.Int32(128)
-            idx_pack = fx.Int32(
-                crd2idx(
-                    [
-                        fx.Int64(n_blk),
-                        fx.Int64(k0),
-                        fx.Int64(lane_div_16),
-                        fx.Int64(n_intra),
-                        fx.Int64(0),
-                    ],
-                    layout_b,
-                )
-            )
-            # idx_pack is a fp4-byte offset; the dwordx4 tile index = (idx_pack/4 dwords)/4.
-            r = fx.make_rmem_tensor(w_reg_lay, fx.Int32)
-            fx.copy(w_copy_atom, fx.slice(w_tiles, (None, idx_pack // fx.Int32(16))), r)
-            v4 = fx.Vector(fx.memref_load_vec(r))
-            raw.append([fx.Int32(v4[j]) for j in range(4)])
-        return raw
-
-    def load_b_raw_int4(base_k, n_blk, n_intra):
-        # kpack=8: K32 block c=k0*4+lane_div_16 = klane slots 2*(c%2),+1 (128B apart) ->
-        # two dwordx2; raw[k0i]=[a#0,a#1,b#0,b#1], each i32 = one K-step.
-        raw = []
-        for k0i in range_constexpr(_k0_count):
-            k0 = (base_k + fx.Int32(k0i * 128)) // fx.Int32(128)
-            c = k0 * fx.Int32(4) + lane_div_16
-            old_k0 = c // fx.Int32(2)
-            old_klane_a = (c % fx.Int32(2)) * fx.Int32(2)
-            four = []
-            for slot in range_constexpr(2):
-                idx = fx.Int32(
-                    crd2idx(
-                        [
-                            fx.Int64(n_blk),
-                            fx.Int64(old_k0),
-                            fx.Int64(old_klane_a + fx.Int32(slot)),
-                            fx.Int64(n_intra),
-                            fx.Int64(0),
-                        ],
-                        layout_b_int4,
-                    )
-                )
-                # idx is an int4-nibble byte offset; dwordx2 tile index = (idx bytes)/8.
-                r = fx.make_rmem_tensor(w_reg_lay2, fx.Int32)
-                fx.copy(
-                    w_copy_atom64, fx.slice(w_tiles8, (None, idx // fx.Int32(8))), r
-                )
-                v2 = fx.Vector(fx.memref_load_vec(r))
-                four.append(fx.Int32(v2[0]))
-                four.append(fx.Int32(v2[1]))
-            raw.append(four)
-        return raw
-
-    def load_b_raw_bf16(base_k, n_blk, n_intra):
-        # Raw bf16 W: one dwordx4 (8 bf16) per ku = one MFMA K32 B fragment (v8bf16,
-        # the MMA operand directly). K map matches fp4: bf_k0 = base_k//32 + (ku//4)*4
-        # + klane_hw, bf_klane = ku%4.
-        raw = []
-        base_k0 = base_k // fx.Int32(32)
-        for ku in range_constexpr(k_unroll):
-            _k0_blk = ku // 4
-            bf_k0 = base_k0 + fx.Int32(_k0_blk * 4) + lane_div_16
-            bf_klane = fx.Int32(ku % 4)
-            elem_idx = fx.Int32(
-                crd2idx(
-                    [
-                        fx.Int64(n_blk),
-                        fx.Int64(bf_k0),
-                        fx.Int64(bf_klane),
-                        fx.Int64(n_intra),
-                        fx.Int64(0),
-                    ],
-                    layout_b_bf16,
-                )
-            )
-            # elem_idx is a bf16-elem offset; dword index = elem_idx*2/4, tile index = /4.
-            r = fx.make_rmem_tensor(w_reg_lay, fx.Int32)
-            fx.copy(w_copy_atom, fx.slice(w_tiles, (None, elem_idx // fx.Int32(8))), r)
-            raw.append(fx.Vector(fx.memref_load_vec(r)).bitcast(fx.BFloat16))  # v8bf16
-        return raw
-
-    def load_b_scale(base_k, mni, n_pack):
-        # aiter _get_scale_f32: adj_ku = base_k//32 + (ku//4)*4 + lane_div_16. Per-lane
-        # scalar e8m0 load on buffer_ops (no layout form, dict-cached across ku).
-        scales = []
-        cache = {}
-        for ku in range_constexpr(k_unroll):
-            _k0_blk = ku // 4
-            adj_ku = base_k // fx.Int32(32) + fx.Int32(_k0_blk * 4) + lane_div_16
-            k_pack_sub = (adj_ku // fx.Int32(4)) % fx.Int32(2)
-            s_ku = adj_ku // fx.Int32(8)
-            if _k0_blk not in cache:
-                idx = (
-                    mni * fx.Int32(sc_stride_n0)
-                    + s_ku * fx.Int32(sc_stride_k0)
-                    + lane_div_16 * fx.Int32(sc_stride_klane)
-                    + lane_mod_16
-                )
-                cache[_k0_blk] = _buffer_i32_scalar_read(sw_tiles, idx, sw_read_atom)
-            packed = cache[_k0_blk]
-            byte_even = k_pack_sub * fx.Int32(2)
-            byte_odd = byte_even + fx.Int32(1)
-            se = _e8m0_byte_to_f32(packed, byte_even)
-            so = _e8m0_byte_to_f32(packed, byte_odd)
-            scales.append((n_pack == fx.Int32(0)).select(se, so))
-        return scales
-
-    def load_b_scale_int4(base_k, n_full):
-        # int4 scale (E,G//2,N,2) bf16: dword = e*(G//2*N) + (adj_ku//2)*N + n_full, half
-        # by parity. N-major, so 16 lanes share one 64B line (vs a G//2-strided gather).
-        scales = []
-        for ku in range_constexpr(k_unroll):
-            _k0_blk = ku // 4
-            adj_ku = base_k // fx.Int32(32) + fx.Int32(_k0_blk * 4) + lane_div_16
-            pair_idx = adj_ku // fx.Int32(2)
-            dword = _scale_expert_base + pair_idx * fx.Int32(N_OUT) + n_full
-            packed = _buffer_i32_scalar_read(sw_tiles, dword, sw_read_atom)
-            # even adj_ku -> low bf16, odd -> high.
-            lo = (packed << fx.Int32(16)).bitcast(fx.Float32)
-            hi = (packed & fx.Int32(0xFFFF0000)).bitcast(fx.Float32)
-            scales.append((adj_ku % fx.Int32(2) == fx.Int32(0)).select(lo, hi))
-        return scales
-
-    vec2_bf16 = T.vec(2, T.bf16)
-
-    def upconvert_b(raw, ku, scale_f32):
-        if const_expr(_is_bf16):
-            # raw[ku] is already the v8bf16 MMA operand (no scale, no upconvert).
-            return raw[ku]
-        i32_val = _raw(raw[ku // 4][ku % 4])
-        if const_expr(_is_int4):
-            return _int4_nibble_to_bf16x8(
-                fx.Int32(i32_val), scale_f32, use_k16=use_k16, old_pack=True
-            )
-        # raw[ku//4][ku%4] i32 holds 8 fp4 -> 4x cvt (v2bf16, sel 0..3) -> v8bf16.
-        s_raw = _raw(scale_f32)
-        i32s = []
-        for sel in range_constexpr(4):
-            p = rocdl.cvt_scalef32_pk_bf16_fp4(vec2_bf16, i32_val, s_raw, sel)
-            i32s.append(fx.Int32(fx.Vector(p).bitcast(fx.Int32)[0]))
-        v4i32 = fx.Vector.from_elements([_raw(x) for x in i32s], fx.Int32)
-        return v4i32.bitcast(fx.BFloat16)  # v8bf16
 
     # ---- N-column addressing for gate/up (SEPARATED; wave owns _n_per_wave) ----
     n_tile_base = wave_n_id * fx.Int32(_n_per_wave)
@@ -521,7 +291,7 @@ def _gemm1_body_a16w4(
             scale_mni_up.append(scale_mni)
             scale_np_up.append(fx.Int32(1))
         else:
-            # bf16 W folds expert_off into the resource base (see w_tiles); mxfp4/int4 index it.
+            # bf16 W folds expert_off into the resource base (utils.make_b_loader); mxfp4/int4 index it.
             _row_expert_off = fx.Int32(0) if const_expr(_is_bf16) else expert_off
             row_gate = _row_expert_off + col_g
             row_up = row_gate + inter_i32
@@ -559,7 +329,7 @@ def _gemm1_body_a16w4(
     else:
         mma_atom = fx.make_mma_atom(fx.rocdl.MFMA(16, 16, 32, fx.BFloat16))
 
-    _mma = functools.partial(mma_bf16, mma_atom, use_k16)
+    _mma = functools.partial(_mma_bf16, mma_atom, use_k16)
 
     # int4 groupwise scale is the OLD-kernel (E, G//2, N, 2) layout: N (col_g | col_g+
     # inter) is WITHIN-expert; the expert base is a separate G//2*N_OUT stride.
@@ -576,11 +346,11 @@ def _gemm1_body_a16w4(
             # Raw bf16 W: no scale; the loaded fragments are the MMA operands.
             return (
                 [
-                    load_b_raw_bf16(base_k, n_blk_gate[ni], n_intra_gate[ni])
+                    b_loader.load_b_raw_bf16(base_k, n_blk_gate[ni], n_intra_gate[ni])
                     for ni in range_constexpr(num_acc_n)
                 ],
                 [
-                    load_b_raw_bf16(base_k, n_blk_up[ni], n_intra_up[ni])
+                    b_loader.load_b_raw_bf16(base_k, n_blk_up[ni], n_intra_up[ni])
                     for ni in range_constexpr(num_acc_n)
                 ],
                 None,
@@ -588,23 +358,23 @@ def _gemm1_body_a16w4(
             )
         if const_expr(_is_int4):
             g_sc = [
-                load_b_scale_int4(base_k, scale_n_gate[ni])
+                b_loader.load_b_scale_int4(base_k, scale_n_gate[ni], _scale_expert_base)
                 for ni in range_constexpr(num_acc_n)
             ]
             u_sc = [
-                load_b_scale_int4(base_k, scale_n_up[ni])
+                b_loader.load_b_scale_int4(base_k, scale_n_up[ni], _scale_expert_base)
                 for ni in range_constexpr(num_acc_n)
             ]
         else:
             g_sc = [
-                load_b_scale(base_k, scale_mni_gate[ni], scale_np_gate[ni])
+                b_loader.load_b_scale(base_k, scale_mni_gate[ni], scale_np_gate[ni])
                 for ni in range_constexpr(num_acc_n)
             ]
             u_sc = [
-                load_b_scale(base_k, scale_mni_up[ni], scale_np_up[ni])
+                b_loader.load_b_scale(base_k, scale_mni_up[ni], scale_np_up[ni])
                 for ni in range_constexpr(num_acc_n)
             ]
-        _lbr = load_b_raw_int4 if const_expr(_is_int4) else load_b_raw
+        _lbr = b_loader.load_raw
         return (
             [
                 _lbr(base_k, n_blk_gate[ni], n_intra_gate[ni])
@@ -623,7 +393,7 @@ def _gemm1_body_a16w4(
         # (aiter phase-separated iteration): drops the per-read vmcnt(0) drains that
         # would otherwise stall the B weight loads.
         return [
-            [lds_load_a(mi, ku, slot=read_slot) for ku in range_constexpr(k_unroll)]
+            [a_loader.load(mi, ku, slot=read_slot) for ku in range_constexpr(k_unroll)]
             for mi in range_constexpr(m_repeat)
         ]
 
@@ -634,8 +404,8 @@ def _gemm1_body_a16w4(
             for ku in range_constexpr(k_unroll):
                 _gsc = None if const_expr(_is_bf16) else g_sc[ni][ku]
                 _usc = None if const_expr(_is_bf16) else u_sc[ni][ku]
-                gb = upconvert_b(g_raw[ni], ku, _gsc)
-                ub = upconvert_b(u_raw[ni], ku, _usc)
+                gb = b_loader.upconvert_b(g_raw[ni], ku, _gsc)
+                ub = b_loader.upconvert_b(u_raw[ni], ku, _usc)
                 for mi in range_constexpr(m_repeat):
                     a8 = a_frags[mi][ku]
                     _mma(acc_gate[mi][ni], a8, gb)
